Made smxi kernel install and smxi tools fully abstracted in terms of kernel handling.
This cleans up a lot of old kanotix era cruft, and helps move everything to be fully future-proofed.

The longer I do smxi, the more I realize ever really worrying about any specific data type is not as good as making it generalized and 
agnostic overall to what data is going through it.

That also makes it a lot easier to maintain long term, fewer gotchas and hard codings that are left unhandled.